Fixed the timer on Linux by using the same code as Android
The fallback for non-Android was broken. It didn't take seconds into
account and it was resetting the starting time at each measurement,
causing the returned time to be the time since the last frame rather
than the amount of seconds since start.

Fortunately the code for Android was doing the right thing and also
works on Linux, so just use that.

For clarity, 'last_time' was renamed to 'start_time'. Also, a redundant
assignment to tv_nsec was removed (it would be reset by clock_gettime
anyway).