add log message MOAI_FileOpenError_S
While this may seem redundant with MOAI_FileNotFound_S, there
may be cases where the file exists but can't be opened, e.g.,
due to permission errors. This also supports a safer code structure:
actually try to open the file and log if it fails, instead of checking
for file existence and then just assuming it will open.