add (untested) support for controlling GL blend equations.
GL_MIN and GL_MAX are omitted because they are absent in iOS; rumor
has it they may exist under the names "GL_MIN_EXT" and "GL_MAX_EXT",
but this seemed harder to accomplish.

The intent is that if you don't use setBlendEquation(), you'll never
see any changes, but it's there as an alternative if you have reason
to prefer subtractive rendering.