Color drawing improvements, plus grid functionality changes
The MOAIGfxDevice draw functionality is extended to allow the caller
to specify a color which is multiplicative with the already computed
"final" color (pen color * ambient color). This is then extended out
through the DrawIndex and Draw calls, although not all objects
support it. The big relevance is that grids are modified to support
the specification of color and scale per-item, so if you want to make
a colorful grid, or tint some portions of the grid (but not necessarily
the entire grid), you can do that. Similarly, items can be scaled, with
the default 1.0 scale being the existing behavior.

This functionality is not yet extended to the serialize/deserialize
part of the grid.

A new sample, grid-color-scale, is added to show how to use this
functionality.