Fix IOS sendMail argument and allocation/crash
The ios sendMail function was using the first Lua argument for
recipient, message and subject.  Updated this to use the correct
arguments.

Also in sendMail, a crash was occuring for me.  This seemed to be due
to an array being alloc'ed but constructed with an autorelease method.
I removed the alloc and just created the array using the autorelease
method and this fixed the crash.