Fixing issues with determining a textbox string bounds.
Currently the GetStringBounds returns a rect which is not correct.
Drawing the textbox with the rect provided doesn't create a textbox
with the text laid out as expected.  Fixing some bugs in the build
layout of text box designer and an issue with MOAIGlyph::GetRect which
was causing these problems.