Updating android host to work after the reorg
The android host was broken after the reorg and was never updated to
fix this.  In the main Android.mk, removed unused uscore, reordered the
static_libs to fix some linking errors with zl and whirlpool.  Also had
to change the typedefs in whirlpool to fix a type conflict with zltypes.

Modified the android host to add similar initializations as the ios
host that are required after the reorg.  Also, something was broken in
MOAIGfxDevice which didn't categorize Android as an OpenGL ES host
which would cause only a black screen to be displayed.

Left to do:  Android host printing is still broken due to some linking
problems.  I may take a look into that next.