Add an optional color parameter to the WriteQuad() routines
Preparation for adding nicer color controls in various
places: We add an optional color parameter to all four WriteQuad()
calls in MOAIGfxDevice, and an alternative TransformAndWriteQuad()
which takes that color and uses it instead of the unmodified
final color. This is always multiplicative; it's mFinalColor *
color. Add the corresponding operator* to ZLColorVec.

Similarly, the Draw() and DrawIndex() functions get an (optional)
additional invocation option which specifies a color. So, for
instance, if MOAIGridDeck2D were to specify a color when calling
this->mDeck->Draw(), the color would propagate down to the final
drawing operation.

Signed-off-by: seebs <seebs@seebs.net>