Create new "fancy" grid type with scale and color features
This patch adds a "MOAIGridFancy" type which allows tinting of grid
items using a palette of colors, as well as variable drawing scale for
each item in the grid. The intent is to allow for the use of tile maps
in cases where some amount of slight modification of tiles (such as dimming
them out, or making them partially transparent) would be desireable,
without the need to duplicate the entire tile set multiple times.

Included also is a trivial sample illustrating how to use the new features.

Signed-off-by: seebs <seebs@seebs.net>