Cleanup/bug fixes, add alpha channel to GridFancy
It is occasionally desireable to be able to control alpha
independently from the rest of the color channels. To this
end, add a separate Alpha field to the GridFancy, make the
alpha component optional (defaulting to 1.0) in SetPalette,
and add utility functions to ZLColorVec to scale the alpha
or color channels by a fixed amount.

Signed-off-by: Seebs <seebs@seebs.net>