*.xcodeproj/project.pbxproj: Consistency in specifying 32-bit builds
Some things (like curl) seem to choke if you allow 64-bit Intel
builds. To fix this, we force 32-bit builds for libmoai stuff,
but we also go through removing all the per-target specifications
of that, and instead specify it at a higher level. This may
or may not actually be correct, but it allowed a build to complete.