Make MOAIParticleSystem:setColor() do something useful
Since a MOAIParticleSystem is already a prop, and thus has a color,
why not use that color (and opacity) as the defaults for particles
which don't have explicit writes to SPRITE_RED, etc.?

So we do two things. First, the sprite defaults change from {1,1,1,1}
to {mR,mG,mB,mA} of the MOAIParticleSystem a script is running on
when it creates the sprite. Second, we add a color() op that lets you
store the current colors of that system in registers. This can be
used in an init script, with set(...SPRITE_RED, r1) calls in the render
script, to cause a sprite to preserve the color the ParticleSystem
had when the init script ran.