pex plugin: Cleanup and bug fixes
This looks larger than it is.

There was a ton of duplicated code between the gravity and radial script
cases, and that's simplified and cleaned up. Also, there's some overly
complicated computations, and dead code. It appears that the intended
angle computations for gravity emitters weren't quite correct. The
sourcePosition and sourcePositionVariance values were ignored. The code
to read in attributes could easily overrun buffers. In theory, all of
these should be fixed now.