proper fix for small texture issues
The chunkSize is only there to define what is the max texture  size to be used when splitting a big texture in fragments.

In case of a smaller image to condition, we use power of 2s, and potentially different width & height.

the loading code has been adapted to take those factors in account.

A new test has been added too demonstrating that it works with a very small non-square texture (64x32).