Introduced a new preprocessor AKU_WITH_IFA which wraps Moai's reference to ASIdentifierManager and the AdSupport framework. It defaults to 0 because having this on by default means that when an iOS app is submitted, the submitter is forced to claim their app is using IFA even if they aren't, and then they are forced to state which "approved" reason they are using IFA for. If the app isn't using IFA, then they can't pass this step. As far as I can tell, Moai is just grabbing this value and exposing it to Lua, which isn't needed if you aren't using it for anything.
I believe this hurdle by Apple is fairly new, and so I added it since I just ran into it while submitted an app that uses Moai. I integrated it in the same way that a module would be, although it technically isn't a module.

Please feel free to refactor this into a nicer form that follows however we plan to handle such things moving forward.