Turned on the Deployment Post Processing Xcode build configuration flag. This actually enables a lot of the code stripping features of Xcode, and in addition, allows other apps which use Moai as a dependency in Xcode to use these features (they will run into compiler errors otherwise). I don't believe this will cause problems for anyone who doesn't care about these features.
Also, I changed the default build configuration for command line builds from Debug to Release. I'm assuming that this wasn't being used by anyone, since "Release" is the more sensible option, and most build scripts override this anyway by passing in their preferred configuration. The reason this feature is important is because when an Xcode project has a custom configuration, then includes a dependent project that does NOT have that same configuration, what does it do? It uses this setting to decide what build configuration to use. So that means that if I make a custom config in my host application called "HockeyApp", for example, then Moai will default to giving me a debug build (which is probably not desirable). In my experience, pretty much every custom config relates to release, and generally you want a release build of your dependencies. So I made this change and I want everyone to be aware of it in case you were relying for some reason the Debug setting....feel free to change it back. My opinion though is that setting it to release is going to work for the vast majority of people more often, and anyone needing debug should come up with a custom solution on their end.