Update AOS NDK build to use gcc (instead of clang)
The latest version of the Android NDK toolchain defaults to clang. Under
clang, freetype fails to build due to an inline-assembly issue. This
issue is corrected (at least for now) by setting NDK_TOOLCHAIN_VERSION
to 4.9, which causes the NDK to use gcc again.

The gcc compiler has apparently been deprecated and will be removed in a
future version. We will need to address this issue again at that time.